refactor: getCategory query
